refactor: structs 폴더 ts 전환 및 타입 추출
