--- src/structs/articlesStructs.js
@@ -1,12 +1,0 @@
-import { coerce, nonempty, nullable, object, partial, string } from 'superstruct';
-import { PageParamsStruct } from './commonStructs.js';
-
-export const GetArticleListParamsStruct = PageParamsStruct;
-
-export const CreateArticleBodyStruct = object({
-  title: coerce(nonempty(string()), string(), (value) => value.trim()),
-  content: nonempty(string()),
-  image: nullable(string()),
-});
-
-export const UpdateArticleBodyStruct = partial(CreateArticleBodyStruct);
