+++ src/structs/authStruct.ts
@@ -1,5 +1,4 @@
-import { coerce, nonempty, nullable, object, partial, string } from 'superstruct';
-import { PageParamsStruct } from './commonStructs.js';
+import { nonempty, object, string, Infer } from 'superstruct';
 
 export const registerBodyStruct = object({
     email: nonempty(string()),
@@ -11,3 +10,6 @@ export const loginBodyStruct = object({
     email: nonempty(string()),
     password: nonempty(string()),
 });
+
+export type RegisterBody = Infer<typeof registerBodyStruct>;
+export type LoginBody = Infer<typeof loginBodyStruct>;
