+++ src/structs/commonStructs.ts
@@ -1,4 +1,5 @@
-import { coerce, integer, object, string, defaulted, optional, enums, nonempty } from 'superstruct';
+import { coerce, integer, object, string, defaulted, optional, enums, nonempty, Infer } from 'superstruct';
+
 
 /** Convert string to integer then validate it */
 const integerString = coerce(integer(), string(), (value) => parseInt(value));
@@ -20,3 +21,7 @@ export const CursorParamsStruct = object({
   orderBy: optional(enums(['recent'])),
   keyword: optional(nonempty(string())),
 });
+
+export type IdParams = Infer<typeof IdParamsStruct>;
+export type PageParams = Infer<typeof PageParamsStruct>;
+export type CursorParams = Infer<typeof CursorParamsStruct>;
