+++ src/structs/userStruct.ts
@@ -1,4 +1,4 @@
-import {object, string, partial, nonempty, nullable} from 'superstruct';
+import {object, string, partial, nonempty, nullable, Infer} from 'superstruct';
 
 export const UpdateUserBodyStruct = partial(
     object({
@@ -11,3 +11,6 @@ export const UpdatePasswordBodyStruct = object({
    oldPassword: nonempty(string()),
    newPassword: nonempty(string()),
 })
+
+export type UpdateUserBody = Infer<typeof UpdateUserBodyStruct>;
+export type UpdatePasswordBody = Infer<typeof UpdatePasswordBodyStruct>;
